+++ src/main/java/school/faang/user_service/mapper/UserMapper.java
@@ -50,13 +50,12 @@ public interface UserMapper {
     @Mapping(target = "participatedEvents", ignore = true)
     @Mapping(target = "country", ignore = true)
     User deactivatedUserDtoToEntity(DeactivatedUserDto deactivatedUserDto);
+
     UserDto toDto(User user);
 
     @Mapping(target = "preference", source = "contactPreference.preference")
     UserForNotificationDto toUserForNotificationDto(User user);
 
-
-    UserForNotificationDto toUserForNotificationDto(User user);
 
     @Named("mapGoalsToListId")
     default List<Long> mapGoalsToListId(List<Goal> goals) {
@@ -85,6 +84,7 @@ public interface UserMapper {
     }
 
     List<User> toEntities(List<UserDto> userDtos);
+
     @Named("mapMentorsToListId")
     default List<Long> mapMentorsToListId(List<User> mentors) {
         if (mentors == null) {
